add common Card component
