+++ src/components/common/card/index.ts
@@ -1,1 +1,2 @@
 export { IngredientCard } from "./ingredient-card"
+export { Card } from "./card"
